Move poster painting into its own module and cover the desktop path
util.js was five unrelated things and is not strict-checked. poster.js takes the
85 lines of poster painting out of it and, being new, goes into CORE_FILES, so the
code that interpolates page-derived URLs into CSS now runs under strict mode with
noUncheckedIndexedAccess.

Adds the assertion the desktop overlay never had: that it paints the thumbnail and
records the outcome its scrim reads. It fails if appendImageAsBackground stops
setting data-thumb-loaded, so the adapter left over from sharing the walk is now
covered rather than merely not crashing.

Fixes a regression from the config-params change: withRemoteConfig walked into
features.duckPlayer.settings.overlays unconditionally, which the minimal configs
do not have, breaking duckplayer-remote-config. It only reaches in when a patch
was actually requested.
