+++ injected/src/features/duckplayer/components/ddg-video-overlay.js
@@ -1,7 +1,7 @@
 import css from '../assets/video-overlay.css';
 import dax from '../assets/dax.svg';
 import { overlayCopyVariants } from '../text.js';
-import { appendImageAsBackground } from '../util.js';
+import { appendImageAsBackground } from '../poster.js';
 import { VideoOverlay } from '../video-overlay.js';
 import { createPolicy, html, trustedUnsafe } from '../../../dom-utils.js';
